Finished Channel CRUD + user chaines improvements

+++ YoutubeClone/Controllers/ChannelController.cs
@@ -34,7 +34,12 @@ namespace YoutubeClone.Controllers
 
         // GET: Default/Details/5
         public ActionResult Details(int id) {
-            var Chaine = db.Chaines.Where(C => C.ChaineId == id).First();
+            var Chaine = db.Chaines.Find(id);
+            if (User.Identity.IsAuthenticated && User.Identity.Name == db.Utilisateurs.Where(c => c.UtilisateurId == db.Chaines.Where(C => C.ChaineId == id).FirstOrDefault().Utilisateur_FK).FirstOrDefault().Username) {
+                ViewBag.EditOK = true;
+            } else {
+                ViewBag.EditOK = false;
+            }
             return View("~/Views/Channel/ChannelViewer.cshtml", Chaine);
         }
 
@@ -80,7 +85,9 @@ namespace YoutubeClone.Controllers
             if (chaine == null) {
                 return HttpNotFound();
             }
-            ViewBag.Utilisateur_FK = new SelectList(db.Utilisateurs, "UtilisateurId", "Username", chaine.Utilisateur_FK);
+
+            var user = db.Utilisateurs.Where(C => C.Username == User.Identity.Name).FirstOrDefault().UtilisateurId;
+            ViewBag.Utilisateur_FK = db.Utilisateurs.Where(C => C.Username == User.Identity.Name).FirstOrDefault().UtilisateurId;
             return View("~/Views/Channel/Edit.cshtml",chaine);
         }
 
@@ -91,12 +98,12 @@ namespace YoutubeClone.Controllers
         [ValidateAntiForgeryToken]
         [Authorize]
         public ActionResult Edit([Bind(Include = "ChaineId,Name,Description,Utilisateur_FK,Categorie_Chaine,Tags_Chaine")] Chaine chaine) {
-            if (ModelState.IsValid) {
-                db.Entry(chaine).State = EntityState.Modified;
+            if (ModelState.IsValid && User.Identity.Name == db.Utilisateurs.Where(c => c.UtilisateurId == db.Chaines.Where(C => C.ChaineId == chaine.ChaineId).FirstOrDefault().Utilisateur_FK).FirstOrDefault().Username) {
+                 db.Entry(chaine).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return View("~/Views/Channel/ChainesUtilisateurs.cshtml", db.Chaines.Where(c => c.Utilisateur_FK == user));
             }
-            ViewBag.Utilisateur_FK = new SelectList(db.Utilisateurs, "UtilisateurId", "Username", chaine.Utilisateur_FK);
+            ViewBag.Utilisateur_FK = db.Utilisateurs.Where(C => C.Username == User.Identity.Name).FirstOrDefault().UtilisateurId;
             return View("~/Views/Channel/Edit.cshtml",chaine);
         }
 
@@ -118,10 +125,13 @@ namespace YoutubeClone.Controllers
         [ValidateAntiForgeryToken]
         [Authorize]
         public ActionResult DeleteConfirmed(int id) {
-            Chaine chaine = db.Chaines.Find(id);
-            db.Chaines.Remove(chaine);
-            db.SaveChanges();
-            return Redirect("Index");
+            if (User.Identity.Name == db.Utilisateurs.Where(c => c.UtilisateurId == db.Chaines.Where(C => C.ChaineId == id).FirstOrDefault().Utilisateur_FK).FirstOrDefault().Username) {
+                Chaine chaine = db.Chaines.Find(id);
+                db.Chaines.Remove(chaine);
+                db.SaveChanges();
+            }
+            var user = db.Utilisateurs.Where(C => C.Username == User.Identity.Name).FirstOrDefault().UtilisateurId;
+            return View("~/Views/Channel/ChainesUtilisateurs.cshtml", db.Chaines.Where(c => c.Utilisateur_FK == user));
         }
 
         protected override void Dispose(bool disposing) {
